% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_riskmetrics.R
\name{add_riskmetrics}
\alias{add_riskmetrics}
\title{Enrich PACKAGES with risk metrics}
\usage{
add_riskmetrics(
  path,
  new_path = path,
  metrics = riskscore::v0.2.1_cran_20230621
)
}
\arguments{
\item{path}{A path to a PACKAGES file.}

\item{new_path}{The path to the risk metrics augmented PACKAGES file. Default
is \code{path}, overwriting.}

\item{metrics}{A data frame providing risk metrics for packages. Packages
indicated in \code{metrics} are matched by package name and version.}
}
\value{
Path of newly created PACKAGES (invisibly).
}
\description{
\code{\link[=add_riskmetrics]{add_riskmetrics()}} adds risk metrics to a PACKAGES file.
}
