% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGES.R
\name{add_score_to_packages}
\alias{add_score_to_packages}
\title{Join PACKAGES meta info with their respective scores (metrics).}
\usage{
add_score_to_packages(packages, scores)
}
\arguments{
\item{packages}{data.frame with the PACKAGES contents.}

\item{scores}{data.frame with the scores.}
}
\value{
data.frame
}
\description{
Join PACKAGES meta info with their respective scores (metrics).
}
