% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CRAN_PACKAGES_FIELDS}
\alias{CRAN_PACKAGES_FIELDS}
\title{The order of PACKAGES fields as in CRAN.}
\format{
An object of class \code{character} of length 16.
}
\usage{
CRAN_PACKAGES_FIELDS
}
\description{
The order of PACKAGES fields as in CRAN.
}
\keyword{internal}
