% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGES.R
\name{diff_packages}
\alias{diff_packages}
\title{Identify new or newer packages.}
\usage{
diff_packages(remote_packages, local_packages)
}
\arguments{
\item{remote_packages}{Data frame with remote packages.}

\item{local_packages}{Data frame with local packages.}
}
\value{
data.frame
}
\description{
New packages - are the packages that are present in the remote repository
but not in the local repository.
}
\details{
Newer packages - are the packages that are present in both repositories,
but the version in the remote repository is newer.
}
