% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{.init_config_values}
\alias{.init_config_values}
\title{Initialize config values from environment variables}
\usage{
.init_config_values()
}
\description{
Initialize config values from environment variables
}
\keyword{internal}
