% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{generate_riskreport}
\alias{generate_riskreport}
\title{Generate a risk report for a package}
\usage{
generate_riskreport(pkg_reference, pkg_assessment)
}
\arguments{
\item{pkg_reference}{output of \code{riskmetric::pkg_ref}}

\item{pkg_assessment}{output of \code{riskmetric::pkg_assess}}
}
\description{
Run riskreport generation, and save the result on disk.
}
