% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{generate_riskreports}
\alias{generate_riskreports}
\title{Generate a risk report for a package}
\usage{
generate_riskreports(
  pkg_reference,
  pkg_assessment,
  output_dir = system.file("report", package = "pharmapkgs")
)
}
\arguments{
\item{pkg_reference}{output of \code{riskmetric::pkg_ref}}

\item{pkg_assessment}{output of \code{riskmetric::pkg_assess}}

\item{output_dir}{where to save the report}
}
\description{
Run riskreport generation, and save the result on disk.
}
\examples{
pkg_ref <- riskmetric::pkg_ref("ggplot2", source = "pkg_cran_remote")
pkg_assessment <- riskmetric::pkg_assess(pkg_ref)
generate_riskreport(pkg_ref, pkg_assessment)

}
