% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGES.R
\name{get_packages}
\alias{get_packages}
\title{Read PACKAGES file from a local or remote repository.}
\usage{
get_packages(
  base_url = .config$remote_base,
  platform = .config$platform,
  r_version = .config$r_version
)
}
\arguments{
\item{base_url}{Base URL of the repository. Can be a local dir path.}

\item{platform}{Platform name.}

\item{r_version}{R version.}
}
\value{
data.frame
}
\description{
The goal of this function is not to produce any significant
side effects, but rather read and return
}
