% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PAK_SPECIAL_FIELDS}
\alias{PAK_SPECIAL_FIELDS}
\title{Special fields that are not present in the PACKAGES file,
but are required for pak to do its magic.}
\format{
An object of class \code{character} of length 1.
}
\usage{
PAK_SPECIAL_FIELDS
}
\description{
Special fields that are not present in the PACKAGES file,
but are required for pak to do its magic.
}
\keyword{internal}
