% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PHARMAPKGS_BASE_URL}
\alias{PHARMAPKGS_BASE_URL}
\title{Base URL for the pharmapkgs repository.}
\format{
An object of class \code{character} of length 1.
}
\usage{
PHARMAPKGS_BASE_URL
}
\description{
Base URL for the pharmapkgs repository.
}
\keyword{internal}
