% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PHARMAPKGS_EXCLUDED_METRICS}
\alias{PHARMAPKGS_EXCLUDED_METRICS}
\title{List of riskmetric assessment names not to run during the scoring.}
\format{
An object of class \code{character} of length 1.
}
\usage{
PHARMAPKGS_EXCLUDED_METRICS
}
\description{
List of riskmetric assessment names not to run during the scoring.
}
\keyword{internal}
