% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{REMOTE_REPO_BASE_URL}
\alias{REMOTE_REPO_BASE_URL}
\title{Base URL for the CRAN repository.}
\format{
An object of class \code{character} of length 1.
}
\usage{
REMOTE_REPO_BASE_URL
}
\description{
Base URL for the CRAN repository.
}
\keyword{internal}
