% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RHUB_BASE_URL}
\alias{RHUB_BASE_URL}
\title{Base URL for the R-hub repository.}
\format{
An object of class \code{character} of length 1.
}
\usage{
RHUB_BASE_URL
}
\description{
Base URL for the R-hub repository.
}
\keyword{internal}
