% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RHUB_PACKAGES_FIELDS}
\alias{RHUB_PACKAGES_FIELDS}
\title{The order of PACKAGES fields as in r-hub/repos.
Notice that some fields are duplicated, this is intentional
and is done to account for different naming conventions that
occurred at different times (e.g. \code{DownloadURL} and \code{DownloadUrl}).}
\format{
An object of class \code{character} of length 25.
}
\usage{
RHUB_PACKAGES_FIELDS
}
\description{
The order of PACKAGES fields as in r-hub/repos.
Notice that some fields are duplicated, this is intentional
and is done to account for different naming conventions that
occurred at different times (e.g. \code{DownloadURL} and \code{DownloadUrl}).
}
\keyword{internal}
