% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RHUB_REPO_PLATFORMS}
\alias{RHUB_REPO_PLATFORMS}
\title{Platforms present in the R-hub repository.}
\format{
An object of class \code{character} of length 9.
}
\usage{
RHUB_REPO_PLATFORMS
}
\description{
Platforms present in the R-hub repository.
}
\keyword{internal}
