% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_filter.R
\name{risk-filter-helpers}
\alias{risk-filter-helpers}
\alias{has_available_packages_field_options}
\alias{risk_filter_id}
\alias{risk_filter_from_conditions}
\alias{risk_filter_with_inserted_fields}
\title{Session uses the "available_packages_field" global option}
\usage{
has_available_packages_field_options()

risk_filter_id()

risk_filter_from_conditions(...)

risk_filter_with_inserted_fields(f, fields)
}
\arguments{
\item{...}{Conditions to use when specifying a risk criteria. Uses
non-standard evaluation, treating risk field names as columns of package
metadata.}

\item{f}{A \code{\link[utils:available.packages]{utils::available.packages()}} filter function}

\item{fields}{A character vector of field names used in the provided \code{...}
conditions. When \code{NULL}, the default, all known risk field names will be
considered.}
}
\description{
Test whether the current session makes use of the
global option, as discussed in the
\href{https://stat.ethz.ch/pipermail/r-devel/2024-June/083477.html}{\code{R-devel} mailing list}.
When used, it is expected that this global option is configured to permit
additional fields to be used by filters. This test can be replaced by an
exact version number once available.
}
