% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_fields.R
\name{risk_fields}
\alias{risk_fields}
\title{Get the risk fields from a repository's RISKFIELDS file.}
\usage{
risk_fields(path)
}
\arguments{
\item{path}{Path to the repository.}
}
\value{
character() with the content of the RISKFIELDS file.
}
\description{
Get the risk fields from a repository's RISKFIELDS file.
}
\examples{
\dontrun{
risk_fields(
  "/home/rmagno/R/x86_64-pc-linux-gnu-library/4.4/pharmapkgs/repos/ubuntu-22.04/4.4/src/contrib"
)
}
}
