% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_filter.R
\name{risk_filter}
\alias{risk_filter}
\title{Create a Risk Filter Function}
\usage{
risk_filter(..., fields = NULL)
}
\arguments{
\item{...}{Conditions to use when specifying a risk criteria. Uses
non-standard evaluation, treating risk field names as columns of package
metadata.}

\item{fields}{A character vector of field names used in the provided \code{...}
conditions. When \code{NULL}, the default, all known risk field names will be
considered.}
}
\description{
Create a risk filter function, with the intended purpose of being provided
to \code{\link[utils:available.packages]{utils::available.packages()}}'s \code{filter} argument, or for use with
the \code{available_packages_filter} option.
}
