% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RUNIVERSE_PACKAGES_FIELDS}
\alias{RUNIVERSE_PACKAGES_FIELDS}
\title{The order of PACKAGES fields as in R-universe.}
\format{
An object of class \code{character} of length 17.
}
\usage{
RUNIVERSE_PACKAGES_FIELDS
}
\description{
The order of PACKAGES fields as in R-universe.
}
\keyword{internal}
