% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGES.R
\name{score_packages}
\alias{score_packages}
\title{Assess packages using riskmetric.}
\usage{
score_packages(packages, limit = .config$limit, repos = .config$remote_base)
}
\arguments{
\item{packages}{Character vector with package names.}

\item{limit}{Maximum number of packages to assess.}
}
\value{
data.frame
}
\description{
Assess packages using riskmetric.
}
