% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_config_value}
\alias{set_config_value}
\title{Set config value}
\usage{
set_config_value(key, value)
}
\arguments{
\item{key}{Name of the config entry.}

\item{value}{Value of the config entry.}
}
\description{
Set config value
}
