% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGES.R
\name{update_packages}
\alias{update_packages}
\title{Update local PACKAGES info.}
\usage{
update_packages(old_local_packages, new_local_packages)
}
\arguments{
\item{old_local_packages}{Data frame with PACKAGES info currently stored in the repo.}

\item{new_local_packages}{Data frame with scored new PACKAGES info.}
}
\value{
data.frame
}
\description{
Update local PACKAGES info.
}
