% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGES.R
\name{write_packages}
\alias{write_packages}
\title{Save PACKAGES info to the local repository.}
\usage{
write_packages(packages, path = .config$local_packages)
}
\arguments{
\item{packages}{Data frame with PACKAGES info.}

\item{path}{Path where to write new PACKAGES file.}
}
\description{
Save PACKAGES info to the local repository.
}
