% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{.config}
\alias{.config}
\title{Config storage to reuse values across the package}
\format{
An object of class \code{environment} of length 4.
}
\usage{
.config
}
\description{
Config storage to reuse values across the package
}
\keyword{internal}
