% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGES.R
\name{get_packages}
\alias{get_packages}
\title{Read PACKAGES file from a local or remote repository.}
\usage{
get_packages(base_url = .config$remote_base)
}
\arguments{
\item{base_url}{Base URL of the repository. Can be a local dir path.}
}
\value{
data.frame
}
\description{
The goal of this function is not to produce any significant
side effects, but rather read and return
}
\details{
Please note, that in order to avoid discrepancies in versioning
semantics, the version field is coerced to \code{base::package_version}.
For example, \code{sf} uses 1.0-19, but the majority of packages would
use 1.0.19. \code{base::package_version("1.0-19")} will return "1.0.19".
Main reason for this is the fact that riskmetric changes the version
of a package from x.y-z to x.y.z, and because of that we can't merge
old existing and new scored packages.
}
