% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PACKAGES.R
\name{get_PACKAGES}
\alias{get_PACKAGES}
\title{Get the PACKAGES file}
\usage{
get_PACKAGES(
  destpath,
  filename = "PACKAGES",
  repos = getOption("repos"),
  type = getOption("pkgType"),
  contrib_url = utils::contrib.url(repos, type),
  quiet = FALSE
)
}
\arguments{
\item{destpath}{Destination path.}

\item{filename}{The name of the PACKAGES file, defaults to... uh...
\code{"PACKAGES"}.}

\item{repos}{Base URL of the repository.}

\item{type}{Package type: \code{"source"}, \code{"mac.binary"}, or \code{"win.binary"}.}

\item{contrib_url}{URL of the contrib sections of the repository.}

\item{quiet}{Passed on to \code{curl::curl_download()}.}
}
\value{
Path of downloaded file (invisibly).
}
\description{
\code{\link[=get_PACKAGES]{get_PACKAGES()}} retrieves the PACKAGES file from a repository.
}
