% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{PHARMAPKGS_BASE_URL}
\alias{PHARMAPKGS_BASE_URL}
\title{Base URL for the pharmapkgs repository.}
\usage{
PHARMAPKGS_BASE_URL()
}
\description{
NOTE: there is an R CMD check issue with a hard-coded path,
that is apparently not available when installing the package
into a temporary directory. Wrapping this in an anonymous
function seems to be a workaround for that issue.
}
\keyword{internal}
