% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_config_value}
\alias{set_config_value}
\title{Set config value}
\usage{
set_config_value(key, value)
}
\arguments{
\item{key}{Name of the config entry.}

\item{value}{Value of the config entry.}
}
\description{
Control how and where the repository and the reports are generated.
}
\details{
Accepted keys are:
\itemize{
\item remote_repo: URL to a remote repository.
\item local_repo: Path to where the repo should be.
\item limit: Number of packages to add.
\item local_packages: Path to where the PACKAGES files is placed.
\item excluded_riskmetric_assessments: Excluded assessments.
}
}
